pyproject.toml
[tool.poetry]
name = "st-demo"
version = "0.1.0"
description = ""
authors = ["weihong.xu <xuweihong.cn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,!=3.9.7"
streamlit = "^1.35.0"
pydantic = "^2.7.4"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
